# perf: Improve resmap performance with AppendAll and Transform functions