--- a/api/internal/builtins/SortOrderTransformer.go
+++ b/api/internal/builtins/SortOrderTransformer.go
@@ -99,6 +99,7 @@ func (p *SortOrderTransformerPlugin) Transform(m resmap.ResMap) (err error) {
 type legacyIDSorter struct {
 	// resids only stores the metadata of the object. This is an optimization as
 	// it's expensive to compute these again and again during ordering.
+
 	resids []resid.ResId
 	// Initially, we sorted the metadata (ResId) of each object and then called GetByCurrentId on each to construct the final list.
 	// The problem is that GetByCurrentId is inefficient and does a linear scan in a list every time we do that.