Personal notes, keeping as draft :)

Rebased, since the SortOrderTransformer PR is in

With the benchmark suite, this PR reduces the runtime from 48s to 2.5s on my machine.
I think we could split it into multiple independent PRs now. I've benchmarked them on their own to see where the major effect 

###    perf: ResAccumulator: Use map for resmap intersection

standalone. 48s to 47s for the benchmark.

### Transform
    perf: NamespaceTransformer: Use Transform function for conflict detection
    perf: resmap: Add Transform function

pair of PRs, 48s to 47s for the benchmark

### AppendMany

commit f5cdc84e6b490a5235f4002efdd5d884310def39
    perf: resmap: Add AppendMany function
no effect on its own

commit e06a96957bb474bbff22ae4940934ff1e8692a47
    perf: improve applyOrdering by avoid call to GetByCurrentId
little effect on its own

commit f6ed5d59f998a8b5f75833faf1fe2e66358f54af
    perf: Improve SortOrderTransformer and use AppendMany
commit 0fe542166a28b7a404a5f4241862d228df6bcb87
    perf: Use AppendMany for SortOrderTransformer/newResMapFromResourceSlice/appendAll
should split this. Here's the major change 48s->2.5s Could split this into more pieces.