@chlunde 2023-11-13
@annasong20 would you like this as one PR, or do you prefer to split some of the commits into different three PRs? I've attempted to make it easier to remove by having a logical story if you start at the first commit, except for [perf: improve applyOrdering by avoid call to GetByCurrentId](https://github.com/kubernetes-sigs/kustomize/pull/5427/commits/36b1715581553c6517308706cb0a73915d0884ae) which is actually another PR that this depends on.