@stormqueen1990 2024-07-15
/remove-lifecycle stale
/auto-cc

On Mon, Jul 15, 2024, 4:23 p.m. Kubernetes Triage Robot <
***@***.***> wrote:

> The Kubernetes project currently lacks enough contributors to adequately
> respond to all PRs.
>
> This bot triages PRs according to the following rules:
>
>    - After 90d of inactivity, lifecycle/stale is applied
>    - After 30d of inactivity since lifecycle/stale was applied,
>    lifecycle/rotten is applied
>    - After 30d of inactivity since lifecycle/rotten was applied, the PR
>    is closed
>
> You can:
>
>    - Mark this PR as fresh with /remove-lifecycle stale
>    - Close this PR with /close
>    - Offer to help out with Issue Triage
>    <https://www.kubernetes.dev/docs/guide/issue-triage/>
>
> Please send feedback to sig-contributor-experience at kubernetes/community
> <https://github.com/kubernetes/community>.
>
> /lifecycle stale
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes-sigs/kustomize/pull/5427#issuecomment-2229335390>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAFKQYME3E5I422CCVKRGXDZMQVVZAVCNFSM6AAAAAA6WWTZC2VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDEMRZGMZTKMZZGA>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>