@koba1t 2024-07-23
@stormqueen1990 
Could I assign you for this and https://github.com/kubernetes-sigs/kustomize/pull/5425 PR?