@koba1t 2024-08-04
I tried to run CI on GitHub Actions.
/assign @stormqueen1990